Directory configuration file, client cache support
Served files list is now in an external file (here=./directory) where also md5 hashes are kept (other type than md5 are planed but not supported/tested yet). If the client calls a request with cache declaration ('If-None-Match') then the '304 Not Modified' response is sent. Red and blue cats for testing. The client MAY support many cache declaration -> untested (all testes browsers declared only one). Loading of the directory is safeguarded with a lock -> preparation for future feature of dynamic reload of the directory.

./updatedir.py - md5 recalculation